+++ src/main/kotlin/de/kaemmelot/datafilesorter/vcs/MissingLineSeparatorException.kt
@@ -1,0 +1,5 @@
+package de.kaemmelot.datafilesorter.vcs
+
+import com.intellij.openapi.vfs.VirtualFile
+
+class MissingLineSeparatorException(val file: VirtualFile) : Exception()